Adds in annotation menu to map points
