+++ TAKAware/Screens/MainScreens/AnnotationDetailView.swift
@@ -1,0 +1,45 @@
+//
+//  AnnotationDetailView.swift
+//  TAKAware
+//
+//  Created by Cory Foy on 7/27/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct AnnotationDetailView: View {
+    @Binding var annotation: MapPointAnnotation?
+    
+    func iconImage() -> UIImage {
+        let icon = IconData.iconFor(type2525: annotation?.cotType ?? "", iconsetPath: annotation?.icon ?? "")
+        return icon.icon
+    }
+
+    var body: some View {
+        VStack(alignment: .leading) {
+            ScrollView {
+                if annotation == nil {
+                    Text("No Map Item Selected")
+                } else {
+                    HStack(alignment: .top) {
+                        VStack {
+                            Group {
+                                Text(annotation!.title ?? "")
+                                Text(annotation!.remarks ?? "")
+                                Text("Type: \(annotation!.cotType ?? "")")
+                                Text("Latitude: \(annotation!.coordinate.latitude.description)")
+                                Text("Longitude: \(annotation!.coordinate.longitude.description)")
+                            }
+                            .frame(maxWidth: .infinity, alignment: .leading)
+                        }
+                        Image(uiImage: iconImage())
+                    }
+                    
+                }
+            }
+        }
+        .padding()
+        .padding(.top, 20)
+    }
+}
