+++ TAKAware/Screens/SettingsScreens/DataPackages.swift
@@ -1,0 +1,15 @@
+//
+//  DataPackages.swift
+//  TAKAware
+//
+//  Created by Cory Foy on 7/26/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct DataPackages: View {
+    var body: some View {
+        Text("Hello, Data Packages!")
+    }
+}
